fixing comments from PR2
