+++ frontend/libs/studio-components/src/components/StudioPageHeader/StudioPageHeader.stories.tsx
@@ -1,35 +1,68 @@
 import React from 'react';
 import type { Meta, StoryFn } from '@storybook/react';
-import { StudioPageHeader, StudioPageHeaderButton } from './index';
+import { StudioPageHeader } from './index';
 import { StudioParagraph } from '../StudioParagraph';
+import { type StudioProfileMenuGroup } from './StudioPageHeaderProfileMenu';
+import { StudioAvatar } from '../StudioAvatar';
+
+const profileMenuGroups: StudioProfileMenuGroup[] = [
+  {
+    items: [
+      {
+        action: { type: 'button', onClick: () => {} },
+        itemName: 'Item 1',
+        isActive: true,
+      },
+      {
+        action: { type: 'link', href: '' },
+        itemName: 'Item 2',
+      },
+    ],
+  },
+  {
+    items: [
+      {
+        action: { type: 'button', onClick: () => {} },
+        itemName: 'Item 3',
+      },
+    ],
+  },
+];
 
 const PreviewComponent = (args): React.ReactElement => (
   <StudioPageHeader {...args}>
     <StudioPageHeader.Main>
       <StudioPageHeader.Left title='Left' showTitle={false} />
       <StudioPageHeader.Center>
-        <StudioPageHeaderButton
+        <StudioPageHeader.HeaderButton
           color={args.variant === 'regular' ? 'light' : 'dark'}
           variant={args.variant}
         >
           Button
-        </StudioPageHeaderButton>
+        </StudioPageHeader.HeaderButton>
       </StudioPageHeader.Center>
       <StudioPageHeader.Right>
-        <StudioParagraph size='sm' style={{ color: 'white', paddingLeft: '20px' }}>
-          Right
-        </StudioParagraph>
+        <div style={{ marginLeft: '10px' }}>
+          <StudioPageHeader.ProfileMenu
+            triggerButtonText='Triggerbutton text'
+            ariaLabelTriggerButton='Triggerbutton aria-label'
+            color='dark'
+            variant='regular'
+            profileMenuGroups={profileMenuGroups}
+            profileImage={<StudioAvatar />}
+          />
+        </div>
       </StudioPageHeader.Right>
     </StudioPageHeader.Main>
     <StudioPageHeader.Sub>
       <div style={{ padding: '10px' }}>
         <StudioParagraph size='sm' style={{ color: 'white' }}>
-          <StudioPageHeaderButton
+          <StudioPageHeader.HeaderButton
             color={args.variant === 'regular' ? 'dark' : 'light'}
             variant={args.variant}
           >
             Subheader Button
-          </StudioPageHeaderButton>
+          </StudioPageHeader.HeaderButton>
         </StudioParagraph>
       </div>
     </StudioPageHeader.Sub>
