+++ frontend/libs/studio-components/src/components/StudioPageHeader/StudioPageHeaderProfileMenu/StudioPageHeaderProfileMenu.module.css
@@ -22,3 +22,7 @@
 .dropDownMenuGroup:not(:first-child) {
   padding-top: var(--fds-spacing-2);
 }
+
+.menuItemButton [aria-checked='true'] {
+  background-color: var(--fds-semantic-surface-neutral-selected);
+}
