+++ frontend/libs/studio-components/src/components/StudioPageHeader/StudioPageHeaderProfileMenu/index.ts
@@ -1,3 +1,6 @@
-export { StudioProfileMenu, type StudioProfileMenuProps } from './StudioProfileMenu';
+export {
+  StudioPageHeaderProfileMenu,
+  type StudioPageHeaderProfileMenuProps,
+} from './StudioPageHeaderProfileMenu';
 export { type StudioProfileMenuItem } from './types/StudioProfileMenuItem';
 export { type StudioProfileMenuGroup } from './types/StudioProfileMenuGroup';
