--- frontend/libs/studio-components/src/components/StudioProfileMenu/components/StudioProfileMenuButton/StudioProfileMenuButton.module.css
@@ -1,3 +1,0 @@
-.selected {
-  background-color: var(--fds-semantic-surface-neutral-selected);
-}
