--- frontend/libs/studio-components/src/components/StudioProfileMenu/components/StudioProfileMenuButton/index.ts
@@ -1,1 +1,0 @@
-export { StudioProfileMenuButton } from './StudioProfileMenuButton';
