--- frontend/libs/studio-components/src/components/StudioProfileMenu/components/StudioProfileMenuLink/index.ts
@@ -1,1 +1,0 @@
-export { StudioProfileMenuLink } from './StudioProfileMenuLink';
